Ensure greater entropy on startup
Previously one of the very first things the program did was assign a
random startup year. In practice this seemed to result in the same year
and same track order being chosen after most cold restarts.

By moving the request for a random value until after the external
hardware has initialised that gives the system plenty of time to collect
sufficient entropy to provide a much less predictable value.